test(client): update snapshots
