+++ packages/client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/test.ts
@@ -41,42 +41,14 @@ test('missing binary, native binaryTarget', async () => {
       await prisma.user.findMany()
     }).rejects.toThrowErrorMatchingInlineSnapshot(`
 
-                                    Invalid \`prisma.user.findMany()\` invocation:
-
-
-                                      Query engine library for current platform "TEST_PLATFORM" could not be found.
-                                    You incorrectly pinned it to TEST_PLATFORM
-
-                                    This probably happens, because you built Prisma Client on a different platform.
-                                    (Prisma Client looked in "/client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/node_modules/@prisma/client/runtime/libquery_engine-TEST_PLATFORM.so.node")
-
-                                    Searched Locations:
-
-                                      /client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/node_modules/.prisma/client
-                                      /client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/node_modules/@prisma/client/runtime
-                                      /client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/node_modules/@prisma/client
-                                      /client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/node_modules/.prisma/client
-                                      /client/src/__tests__/integration/errors/missing-binary-native-binaryTarget
-                                      /tmp/prisma-engines
-                                      /client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/node_modules/.prisma/client
-
-                                    You already added the platform "native" to the "generator" block
-                                    in the "schema.prisma" file as described in https://pris.ly/d/client-generator,
-                                    but something went wrong. That's suboptimal.
-
-                                    Please create an issue at TEST_GITHUB_LINK
-                              `)
-  } else {
-    await expect(async () => {
-      await prisma.user.findMany()
-    }).rejects.toThrowErrorMatchingInlineSnapshot(`
-
             Invalid \`prisma.user.findMany()\` invocation:
 
 
-              Query engine binary for current platform "TEST_PLATFORM" could not be found.
+              Query engine library for current platform "TEST_PLATFORM" could not be found.
+            You incorrectly pinned it to TEST_PLATFORM
+
             This probably happens, because you built Prisma Client on a different platform.
-            (Prisma Client looked in "/client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/node_modules/@prisma/client/runtime/query-engine-TEST_PLATFORM")
+            (Prisma Client looked in "/client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/node_modules/@prisma/client/runtime/libquery_engine-TEST_PLATFORM.dylib.node")
 
             Searched Locations:
 
@@ -94,5 +66,33 @@ test('missing binary, native binaryTarget', async () => {
 
             Please create an issue at TEST_GITHUB_LINK
           `)
+  } else {
+    await expect(async () => {
+      await prisma.user.findMany()
+    }).rejects.toThrowErrorMatchingInlineSnapshot(`
+
+                        Invalid \`prisma.user.findMany()\` invocation:
+
+
+                          Query engine binary for current platform "TEST_PLATFORM" could not be found.
+                        This probably happens, because you built Prisma Client on a different platform.
+                        (Prisma Client looked in "/client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/node_modules/@prisma/client/runtime/query-engine-TEST_PLATFORM")
+
+                        Searched Locations:
+
+                          /client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/node_modules/.prisma/client
+                          /client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/node_modules/@prisma/client/runtime
+                          /client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/node_modules/@prisma/client
+                          /client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/node_modules/.prisma/client
+                          /client/src/__tests__/integration/errors/missing-binary-native-binaryTarget
+                          /tmp/prisma-engines
+                          /client/src/__tests__/integration/errors/missing-binary-native-binaryTarget/node_modules/.prisma/client
+
+                        You already added the platform "native" to the "generator" block
+                        in the "schema.prisma" file as described in https://pris.ly/d/client-generator,
+                        but something went wrong. That's suboptimal.
+
+                        Please create an issue at TEST_GITHUB_LINK
+                    `)
   }
 })
