+++ packages/client/src/__tests__/integration/errors/missing-binary/test.ts
@@ -42,46 +42,14 @@ test('missing-binary', async () => {
       await prisma.user.findMany()
     }).rejects.toThrowErrorMatchingInlineSnapshot(`
 
-                                                                        Invalid \`prisma.user.findMany()\` invocation:
-
-
-                                                                          Query engine library for current platform "TEST_PLATFORM" could not be found.
-                                                                        You incorrectly pinned it to TEST_PLATFORM
-
-                                                                        This probably happens, because you built Prisma Client on a different platform.
-                                                                        (Prisma Client looked in "/client/src/__tests__/integration/errors/missing-binary/node_modules/@prisma/client/runtime/libquery_engine-TEST_PLATFORM.so.node")
-
-                                                                        Searched Locations:
-
-                                                                          /client/src/__tests__/integration/errors/missing-binary/node_modules/.prisma/client
-                                                                          /client/src/__tests__/integration/errors/missing-binary/node_modules/@prisma/client/runtime
-                                                                          /client/src/__tests__/integration/errors/missing-binary/node_modules/@prisma/client
-                                                                          /client/src/__tests__/integration/errors/missing-binary/node_modules/.prisma/client
-                                                                          /client/src/__tests__/integration/errors/missing-binary
-                                                                          /tmp/prisma-engines
-                                                                          /client/src/__tests__/integration/errors/missing-binary/node_modules/.prisma/client
-
-
-                                                                        To solve this problem, add the platform "TEST_PLATFORM" to the "binaryTargets" attribute in the "generator" block in the "schema.prisma" file:
-                                                                        generator client {
-                                                                          provider      = "prisma-client-js"
-                                                                          binaryTargets = ["native"]
-                                                                        }
-
-                                                                        Then run "prisma generate" for your changes to take effect.
-                                                                        Read more about deploying Prisma Client: https://pris.ly/d/client-generator
-                                                            `)
-  } else {
-    await expect(async () => {
-      await prisma.user.findMany()
-    }).rejects.toThrowErrorMatchingInlineSnapshot(`
-
             Invalid \`prisma.user.findMany()\` invocation:
 
 
-              Query engine binary for current platform "TEST_PLATFORM" could not be found.
+              Query engine library for current platform "TEST_PLATFORM" could not be found.
+            You incorrectly pinned it to TEST_PLATFORM
+
             This probably happens, because you built Prisma Client on a different platform.
-            (Prisma Client looked in "/client/src/__tests__/integration/errors/missing-binary/node_modules/@prisma/client/runtime/query-engine-TEST_PLATFORM")
+            (Prisma Client looked in "/client/src/__tests__/integration/errors/missing-binary/node_modules/@prisma/client/runtime/libquery_engine-TEST_PLATFORM.dylib.node")
 
             Searched Locations:
 
@@ -103,5 +71,37 @@ test('missing-binary', async () => {
             Then run "prisma generate" for your changes to take effect.
             Read more about deploying Prisma Client: https://pris.ly/d/client-generator
           `)
+  } else {
+    await expect(async () => {
+      await prisma.user.findMany()
+    }).rejects.toThrowErrorMatchingInlineSnapshot(`
+
+                        Invalid \`prisma.user.findMany()\` invocation:
+
+
+                          Query engine binary for current platform "TEST_PLATFORM" could not be found.
+                        This probably happens, because you built Prisma Client on a different platform.
+                        (Prisma Client looked in "/client/src/__tests__/integration/errors/missing-binary/node_modules/@prisma/client/runtime/query-engine-TEST_PLATFORM")
+
+                        Searched Locations:
+
+                          /client/src/__tests__/integration/errors/missing-binary/node_modules/.prisma/client
+                          /client/src/__tests__/integration/errors/missing-binary/node_modules/@prisma/client/runtime
+                          /client/src/__tests__/integration/errors/missing-binary/node_modules/@prisma/client
+                          /client/src/__tests__/integration/errors/missing-binary/node_modules/.prisma/client
+                          /client/src/__tests__/integration/errors/missing-binary
+                          /tmp/prisma-engines
+                          /client/src/__tests__/integration/errors/missing-binary/node_modules/.prisma/client
+
+
+                        To solve this problem, add the platform "TEST_PLATFORM" to the "binaryTargets" attribute in the "generator" block in the "schema.prisma" file:
+                        generator client {
+                          provider      = "prisma-client-js"
+                          binaryTargets = ["native"]
+                        }
+
+                        Then run "prisma generate" for your changes to take effect.
+                        Read more about deploying Prisma Client: https://pris.ly/d/client-generator
+                    `)
   }
 })
