+++ packages/client/src/__tests__/integration/errors/referentialActions-onDelete-default-foreign-key-error-mysql/schema.prisma
