+++ packages/client/src/__tests__/integration/errors/referentialActions-onDelete-default-foreign-key-error-mysql/test.ts
@@ -1,12 +1,17 @@
 import path from 'path'
 import { generateTestClient } from '../../../../utils/getTestClient'
+import { tearDownMysql } from '../../../../utils/setupMysql'
 import { migrateDb } from '../../__helpers__/migrateDb'
 
 let prisma
-describe('default-onDelete-cascade(sqlite)', () => {
+const baseUri = process.env.TEST_MYSQL_URI
+
+describe('referentialActions-onDelete-default-foreign-key-error(mysql)', () => {
   beforeAll(async () => {
+    process.env.TEST_MYSQL_URI += '-default-onDelete-Cascade'
+    await tearDownMysql(process.env.TEST_MYSQL_URI!)
     await migrateDb({
-      connectionString: `file:./dev.db`,
+      connectionString: process.env.TEST_MYSQL_URI!,
       schemaPath: path.join(__dirname, 'schema.prisma'),
     })
     await generateTestClient()
@@ -19,6 +24,7 @@ describe('default-onDelete-cascade(sqlite)', () => {
     await prisma.profile.deleteMany()
     await prisma.user.deleteMany()
     await prisma.$disconnect()
+    process.env.TEST_MYSQL_URI = baseUri
   })
 
   test('delete 1 user, should error', async () => {
@@ -47,13 +53,13 @@ describe('default-onDelete-cascade(sqlite)', () => {
       expect(e.message).toMatchInlineSnapshot(`
 
 Invalid \`prisma.user.delete()\` invocation in
-/client/src/__tests__/integration/errors/default-onDelete-cascade-sqlite/test.ts:41:31
+/client/src/__tests__/integration/errors/default-onDelete-cascade-mysql/test.ts:47:31
 
-  38 expect(await prisma.user.findMany()).toHaveLength(1)
-  39 
-  40 try {
-→ 41   await prisma.user.delete(
-  The change you are trying to make would violate the required relation 'PostToUser' between the \`Post\` and \`User\` models.
+  44 expect(await prisma.user.findMany()).toHaveLength(1)
+  45 
+  46 try {
+→ 47   await prisma.user.delete(
+  Foreign key constraint failed on the field: \`authorId\`
 `)
     }
   })
