+++ packages/client/src/__tests__/integration/errors/referentialActions-onDelete-default-foreign-key-error-postgresql/test.ts
@@ -1,14 +1,17 @@
 import path from 'path'
 import { generateTestClient } from '../../../../utils/getTestClient'
+import { tearDownPostgres } from '../../../../utils/setupPostgres'
 import { migrateDb } from '../../__helpers__/migrateDb'
 
 let prisma
-// skipped because flaky https://buildkite.com/prisma/prisma2-publish/builds/4902#c94c9d75-8d51-4abe-a875-13fd2a4ee6fe/179-1114
-// errors with: `The table `dbo.UserDefaultOnDelete` does not exist in the current database.`
-describe.skip('referentialActions(sqlserver)', () => {
+const baseUri = process.env.TEST_POSTGRES_URI
+
+describe('referentialActions-onDelete-default-foreign-key-error(postgresql)', () => {
   beforeAll(async () => {
+    process.env.TEST_POSTGRES_URI += '-default-onDelete-Cascade'
+    await tearDownPostgres(process.env.TEST_POSTGRES_URI!)
     await migrateDb({
-      connectionString: process.env.TEST_MSSQL_JDBC_URI!,
+      connectionString: process.env.TEST_POSTGRES_URI!,
       schemaPath: path.join(__dirname, 'schema.prisma'),
     })
     await generateTestClient()
@@ -21,6 +24,7 @@ describe.skip('referentialActions(sqlserver)', () => {
     await prisma.profile.deleteMany()
     await prisma.user.deleteMany()
     await prisma.$disconnect()
+    process.env.TEST_POSTGRES_URI = baseUri
   })
 
   test('delete 1 user, should error', async () => {
@@ -49,13 +53,13 @@ describe.skip('referentialActions(sqlserver)', () => {
       expect(e.message).toMatchInlineSnapshot(`
 
 Invalid \`prisma.user.delete()\` invocation in
-/client/src/__tests__/integration/errors/default-onDelete-cascade-sqlserver/test.ts:41:31
+/client/src/__tests__/integration/errors/default-onDelete-cascade-postgres/test.ts:47:31
 
-  38 expect(await prisma.user.findMany()).toHaveLength(1)
-  39 
-  40 try {
-→ 41   await prisma.user.delete(
-  The change you are trying to make would violate the required relation 'PostToUser' between the \`Post\` and \`User\` models.
+  44 expect(await prisma.user.findMany()).toHaveLength(1)
+  45 
+  46 try {
+→ 47   await prisma.user.delete(
+  Foreign key constraint failed on the field: \`Post_authorId_fkey (index)\`
 `)
     }
   })
