+++ packages/client/src/__tests__/integration/errors/referentialActions-onDelete-default-foreign-key-error-sqlserver/test.ts
@@ -1,17 +1,15 @@
 import path from 'path'
 import { generateTestClient } from '../../../../utils/getTestClient'
-import { tearDownPostgres } from '../../../../utils/setupPostgres'
 import { migrateDb } from '../../__helpers__/migrateDb'
 
 let prisma
-const baseUri = process.env.TEST_POSTGRES_URI
-
-describe('default-onDelete-cascade(postgresql)', () => {
+// skipped because flaky https://buildkite.com/prisma/prisma2-publish/builds/4902#c94c9d75-8d51-4abe-a875-13fd2a4ee6fe/179-1114
+// errors with: `The table `dbo.UserDefaultOnDelete` does not exist in the current database.`
+// TODO unskip
+describe.skip('referentialActions-onDelete-default-foreign-key-error(sqlserver)', () => {
   beforeAll(async () => {
-    process.env.TEST_POSTGRES_URI += '-default-onDelete-Cascade'
-    await tearDownPostgres(process.env.TEST_POSTGRES_URI!)
     await migrateDb({
-      connectionString: process.env.TEST_POSTGRES_URI!,
+      connectionString: process.env.TEST_MSSQL_JDBC_URI!,
       schemaPath: path.join(__dirname, 'schema.prisma'),
     })
     await generateTestClient()
@@ -24,7 +22,6 @@ describe('default-onDelete-cascade(postgresql)', () => {
     await prisma.profile.deleteMany()
     await prisma.user.deleteMany()
     await prisma.$disconnect()
-    process.env.TEST_POSTGRES_URI = baseUri
   })
 
   test('delete 1 user, should error', async () => {
@@ -53,12 +50,12 @@ describe('default-onDelete-cascade(postgresql)', () => {
       expect(e.message).toMatchInlineSnapshot(`
 
 Invalid \`prisma.user.delete()\` invocation in
-/client/src/__tests__/integration/errors/default-onDelete-cascade-postgres/test.ts:47:31
+/client/src/__tests__/integration/errors/default-onDelete-cascade-sqlserver/test.ts:41:31
 
-  44 expect(await prisma.user.findMany()).toHaveLength(1)
-  45 
-  46 try {
-→ 47   await prisma.user.delete(
+  38 expect(await prisma.user.findMany()).toHaveLength(1)
+  39 
+  40 try {
+→ 41   await prisma.user.delete(
   The change you are trying to make would violate the required relation 'PostToUser' between the \`Post\` and \`User\` models.
 `)
     }
