+++ packages/client/src/__tests__/integration/happy/referentialActions-onDelete-cascade-mysql/test.ts
@@ -64,7 +64,5 @@ describe('referentialActions(mysql)', () => {
     expect(await prisma.user.findMany()).toHaveLength(1)
     expect(await prisma.profile.findMany()).toHaveLength(1)
     expect(await prisma.post.findMany()).toHaveLength(1)
-
-    // console.debug({ deleteBob })
   })
 })
