+++ packages/client/src/__tests__/integration/happy/referentialActions-onDelete-cascade-postgresql/schema.prisma
@@ -4,8 +4,7 @@ datasource db {
 }
 
 generator client {
-  provider        = "prisma-client-js"
-  previewFeatures = ["referentialActions"]
+  provider = "prisma-client-js"
 }
 
 model Post {
