+++ packages/client/src/__tests__/integration/happy/referentialActions-onDelete-cascade-sqlserver/schema.prisma
@@ -5,7 +5,7 @@ datasource my_db {
 
 generator client {
   provider        = "prisma-client-js"
-  previewFeatures = ["microsoftSqlServer", "referentialActions"]
+  previewFeatures = ["microsoftSqlServer"]
 }
 
 model Post {
