+++ packages/client/src/__tests__/integration/happy/referentialActions-onDelete-cascade-sqlserver/test.ts
@@ -58,7 +58,5 @@ describe('referentialActions(postgresql)', () => {
     expect(await prisma.user.findMany()).toHaveLength(1)
     expect(await prisma.profile.findMany()).toHaveLength(1)
     expect(await prisma.post.findMany()).toHaveLength(1)
-
-    // console.debug({ deleteBob })
   })
 })
